+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/CompilerBeanEnactor2.java
@@ -20,8 +20,7 @@ import java.util.List;
 import java.util.function.Supplier;
 
 import static org.openprovenance.prov.template.compiler.ConfigProcessor.*;
-import static org.openprovenance.prov.template.compiler.configuration.SpecificationFile.generateJava;
-import static org.openprovenance.prov.template.compiler.configuration.SpecificationFile.generatePython;
+import static org.openprovenance.prov.template.compiler.configuration.SpecificationFile.*;
 import static org.openprovenance.prov.template.compiler.past.Assignment.ASSIGNMENT;
 import static org.openprovenance.prov.template.compiler.past.Constructor.CONSTRUCTOR;
 import static org.openprovenance.prov.template.compiler.past.Field.FIELD;
@@ -135,8 +134,9 @@ public class CompilerBeanEnactor2 {
 
         Supplier<Boolean> pythonGenerator=() -> generatePython(pastClass, myPackage, locations.python_dir, stackTraceElement);
         Supplier<Boolean> javaGenerator = () -> generateJava(pastClass, myPackage, configs, fileName + ".java", directory, stackTraceElement, compilerUtil);
+        Supplier<Boolean> jsGenerator = () -> generateJavaScript(pastClass, myPackage, "target/generated-js", stackTraceElement);
 
-        return new SpecificationFile(javaGenerator,pythonGenerator);
+        return new SpecificationFile(javaGenerator,pythonGenerator, javaGenerator,jsGenerator);
 
     }
 
@@ -180,8 +180,9 @@ public class CompilerBeanEnactor2 {
 
         Supplier<Boolean> pythonGenerator=() -> generatePython(intfce, myPackage, locations.python_dir, stackTraceElement);
         Supplier<Boolean> javaGenerator = () -> generateJava(intfce, myPackage, configs, fileName + ".java", directory, stackTraceElement, compilerUtil);
+        Supplier<Boolean> jsGenerator = () -> generateJavaScript(intfce, myPackage, "target/generated-js", stackTraceElement);
 
-        return new SpecificationFile(javaGenerator,pythonGenerator);
+        return new SpecificationFile(javaGenerator,pythonGenerator,  jsGenerator, emptyGenerator);
 
     }