+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/annotations/AnnotationConverter.java
@@ -14,6 +14,7 @@ public class AnnotationConverter {
         map.put(RegisterMethod.NAME, new RegisterMethod());
         map.put(SingleDispatchMethod.NAME, new SingleDispatchMethod());
         map.put(OverrideAnnotation.NAME, new OverrideAnnotation());
+        map.put(StaticMethod.NAME, new StaticMethod());
         return map;
     }