+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/annotations/JsAnnotation.java
@@ -1,0 +1,21 @@
+package org.openprovenance.prov.template.compiler.past.annotations;
+
+abstract public class JsAnnotation implements PastAnnotation {
+    private final String name;
+
+    public JsAnnotation(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public String getLanguage() {
+        return "JS";
+    }
+
+
+}