+++ RELEASES.md
@@ -2,6 +2,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.2.4-SNAPSHOT`
 * prov-template-compiler: StatementCompilerAction2: refined past code generator for Attribution and Association
+* prov-template-compiler: javascript emitter updates to run workflow in js
 
 `2.2.3`
 * prov-template: update to instantiation